Fixed bug where closed tree windows were still being saved in projects
git-svn-id: svn+ssh://svn.code.sf.net/p/topiaryexplorer/code/trunk/main@396 05d1262b-b726-4fff-8b9f-3609a8a82e43

+++ src/topiaryexplorer/TreeWindow.java
@@ -1306,7 +1306,7 @@ public class TreeWindow extends TopiaryWindow implements KeyListener, ActionList
              if (tree.getTree() == null) return;
              double value = treeEditToolbar.branchEditPanel.lineWidthSlider.getValue();
              value = value/20.0;
-             tree.setLineWidthScale(value);
+             tree.setLineWidthScale((float)value);
              tree.redraw();
          }
